Revert "feat: course recommendations for Skills Builder"
