+++ src/skills-builder/skills-builder-modal/select-preferences/CareerInterestSelect.jsx
@@ -1,9 +1,7 @@
 import React, { useContext } from 'react';
 import { getConfig } from '@edx/frontend-platform';
 import { useIntl } from '@edx/frontend-platform/i18n';
-import {
-  Stack, Row, Col, Form,
-} from '@edx/paragon';
+import { Stack, Row, Col } from '@edx/paragon';
 import { InstantSearch } from 'react-instantsearch-hooks-web';
 import JobTitleInstantSearch from './JobTitleInstantSearch';
 import CareerInterestCard from './CareerInterestCard';
@@ -18,24 +16,25 @@ const CareerInterestSelect = () => {
   const { searchClient } = algolia;
 
   const handleCareerInterestSelect = (value) => {
-    if (!careerInterests.includes(value) && careerInterests.length < 3) {
+  // By checking for a value to exist, we avoid adding a null value to the careerInterests array
+  // The 'onSelected' function is fired during every 'onChange' event
+  // A null value was being passed to this function whenever the search box received input, resulting in empty cards
+    if (value && careerInterests.length < 3) {
       dispatch(addCareerInterest(value));
     }
   };
 
   return (
     <Stack gap={2}>
-      <Form.Label>
-        <h4 className="mb-3">
-          {formatMessage(messages.careerInterestPrompt)}
-        </h4>
-        <InstantSearch searchClient={searchClient} indexName={getConfig().ALGOLIA_JOBS_INDEX_NAME}>
-          <JobTitleInstantSearch
-            onSelected={handleCareerInterestSelect}
-            placeholder={formatMessage(messages.careerInterestInputPlaceholderText)}
-          />
-        </InstantSearch>
-      </Form.Label>
+      <h4>
+        {formatMessage(messages.careerInterestPrompt)}
+      </h4>
+      <InstantSearch searchClient={searchClient} indexName={getConfig().ALGOLIA_JOBS_INDEX_NAME}>
+        <JobTitleInstantSearch
+          onSelected={handleCareerInterestSelect}
+          placeholder={formatMessage(messages.careerInterestInputPlaceholder)}
+        />
+      </InstantSearch>
       <Row>
         {careerInterests.map((interest, index) => (
           // eslint-disable-next-line react/no-array-index-key
