--- src/skills-builder/skills-builder-modal/view-results/data/constants.js
@@ -1,11 +1,0 @@
-const COURSE = 'course';
-
-/* The below strings can be used to demonstrate how we are able to retrieve recommendations for other product types
-const BOOT_CAMP = 'boot camp';
-const EXECUTIVE_EDUCATION = 'executive education';
-*/
-
-// eslint-disable-next-line import/prefer-default-export
-export const productTypes = [
-  COURSE,
-];
