+++ src/skills-builder/skills-builder-modal/view-results/messages.js
@@ -6,11 +6,6 @@ const messages = defineMessages({
     defaultMessage: 'We found skills and courses that match your preferences!',
     description: 'Success alert message to display when recommendations are presented to the learner.',
   },
-  matchesNotFoundDangerAlert: {
-    id: 'matches.not.found.danger.alert',
-    defaultMessage: 'We were not able to retrieve recommendations at this time. Please try again later.',
-    description: 'Danger alert message to display when the component fails to get recommendations.',
-  },
   relatedSkillsHeading: {
     id: 'related.skills.heading',
     defaultMessage: 'Related Skills',
@@ -20,11 +15,6 @@ const messages = defineMessages({
     id: 'related.skills.selectable.box.label.text',
     defaultMessage: 'Related skills:',
     description: 'Label text for a selectable box that displays related skills for a corresponding selected job title.',
-  },
-  productRecommendationsHeaderText: {
-    id: 'product.recommendations.header.text',
-    defaultMessage: '{productType} recommendations for {jobName}',
-    description: 'Header text for a carousel of product recommendations.',
   },
 });
