+++ src/skills-builder/utils/tests/search.test.jsx
@@ -55,7 +55,7 @@ describe('Algolias utility function', () => {
 
   it('getProductRecommendations() queries Algolia with the expected search parameters', async () => {
     const expectedSearchParameters = {
-      filters: 'product: "Course"',
+      filters: 'product:Course',
       facetFilters: [
         ['skills.skill:Sword Lobbing'],
       ],
